Split ActionTesting.hpp into four files.
No actual changes to any code, except for moving blocks of code from
one file to another, clang-formatting, and changes in includes and
forward declarations that were necessary when code blocks were moved.

New files:
 - MockRuntimeSystem.hpp contains the MockRuntimeSystem class.
 - MockRuntimeSystemFreeFunctions.hpp contains the free functions
   that are used to invoke member functions of MockRuntimeSystem
   and a few other free functions that depend on MockRuntimeSystem.
 - MockDistributedObject.hpp contains the MockDistributedObject class.
